# mobile/openapi/identity-service.yaml
...
openapi: 3.1.0
info:
  title: TD IT Solutions — Identity Service API
  version: "1.0.0"
  description: >
    Feature 001 (Customer Account Creation & Authentication). Mediates all
    client access to Supabase Auth per ADR-0002's mediation principle —
    no client ever holds a Supabase-honoured credential (see api-design.md §1).
servers:
  - url: /v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Backend-minted access token (api-design.md §1). Never a Supabase-issued JWT.
    internalServiceAuth:
      type: apiKey
      in: header
      name: X-Internal-Service-Key
      description: Service-to-service auth for the internal-only account-status surface (§2.3 row 4). Not reachable by any client.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Client-generated UUID v4. See api-design.md §4.
...
  headers:
...
    RetryAfter:
      schema: { type: integer, description: Seconds. }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            requestId: { type: string, format: uuid }
...
    SessionTokens:
      type: object
      required: [accessToken, refreshToken, expiresIn, sessionId]
      properties:
        accessToken: { type: string, description: Backend-minted JWT, api-design.md §1. }
        refreshToken: { type: string }
        expiresIn: { type: integer, example: 600, description: Seconds (10 min ceiling, api-design.md §2). }
        sessionId: { type: string, format: uuid }
    MfaChallengeRequired:
      type: object
      required: [mfaRequired, mfaChallengeToken, expiresIn]
      properties:
        mfaRequired: { type: boolean, enum: [true] }
        mfaChallengeToken: { type: string }
        expiresIn: { type: integer, example: 300 }
...
  responses:
...
    Unauthorized:
      description: Missing/invalid/expired/revoked bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Idempotency-key reuse with a different body, or a state conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Locked:
      description: Account temporarily locked (rate-limit escalation, api-design.md §5).
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UpstreamUnavailable:
      description: Supabase Admin API unreachable/timed out (api-design.md §6).
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /invitations/{token}/accept:
    post:
      operationId: acceptInvitation
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: >
            Account created, invitation marked accepted. Does NOT issue a
            session — the caller must proceed to mandatory MFA enrollment
            (BR-4) before any session is minted; see /mfa/enroll.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId: { type: string, format: uuid }
                  mfaEnrollmentRequired: { type: boolean, enum: [true] }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410':
          description: Invitation expired or already accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }

  # ---------------------------------------------------------------
  # MFA enrollment
  # ---------------------------------------------------------------
  /mfa/enroll:
    post:
      operationId: mfaEnroll
      summary: Starts TOTP enrollment. Callable by an unauthenticated-but-invitation-verified privileged user (first login) or an authenticated customer opting in.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId:
                  type: string
                  format: uuid
                  description: Required only for the forced-first-login (post-accept, pre-session) path; omitted when called with a bearer token.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  qrCodeImage: { type: string, description: Base64-encoded PNG. }
                  manualEntryKey: { type: string }
                  enrollmentId: { type: string, format: uuid }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }
  /mfa/enroll/verify:
    post:
      operationId: mfaEnrollVerify
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollmentId, code]
              properties:
                enrollmentId: { type: string, format: uuid }
                code: { type: string, minLength: 6, maxLength: 6 }
      responses:
        '200':
          description: Enrollment confirmed. For the forced-first-login path, this now issues session tokens (mandatory MFA gate cleared).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTokens' }
        '400':
          description: Wrong/expired code. No attempt-count lockout at enrollment (ui-design.md §4.4).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  # ---------------------------------------------------------------
  # Login & session issuance
  # ---------------------------------------------------------------
  /auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                deviceId:
                  type: string
                  nullable: true
                  description: Mobile clients only (FR-20). See api-design.md §3.1 / FU-09.
      responses:
        '200':
          description: Either a full session (no MFA required/enabled) or an MFA challenge.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SessionTokens'
                  - $ref: '#/components/schemas/MfaChallengeRequired'
        '401':
          description: >
            Generic "incorrect email or password" (AC-5) — never distinguishes
            unknown-email from wrong-password. Body includes attemptsRemaining
            when ≤1 (api-design.md §5 escalating-warning contract).
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Error' }
                  - type: object
                    properties:
                      attemptsRemaining: { type: integer, nullable: true }
        '423': { $ref: '#/components/responses/Locked' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }
...